feat: Support secure and insecure client creation method
Previously there is only one createClientOptions method to prepare the xmlprc client instance. To
also support the secure client, the createClientOptions method is splitted into two methods, the
createSecureClientOptions is introduced to create the secure client along with
createInsecureClientOptions to create the insecure client.

BREAKING CHANGE: The createClientOptions is no longer exist and replaced by
createSecureClientOptions and createInsecureClientOptions. As those method returns a maybe, the
execute(Un|A)uthenticatedClient also needs to be refactored to accept maybe instead.